requirements.txt
requests==2.31.0
python-dotenv==1.0.1
flask==3.0.2
gunicorn==20.1.0
